Disabled auth requirements for queries in dev
